+++ components/Header/Header.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link';
 import { ROUTES } from '@/utils/routes';
 import { MirageGalleryLogo } from '../Svgs/MirageGalleryLogo';
-import ConnectWalletButton from '@/components/ConnectWalletButton/ConnectWalletButton';
+import ConnectWalletButton from '@/components/Button/ConnectWalletButton/ConnectWalletButton';
 export default function Header() {
   return (
     <header className="fixed top-0 left-0 right-0 z-50 px-2 py-6 bg-white border-b border-gray-200">
